backend/requirements.txt
django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.3 